v4l2object: Improve colorspace handling to support uvc device
v4l2 encoder won't probe all colorspace, range, matrix and transfer combination
to save initialization time for low-speed devices. Unfortunately uvc driver
default colorimetry is not in the colorimetry list of v4l2 encoder and it causes
not-negotiated. If desired colorspace is 1:4:7:1, uvc driver will produce 1:4:5:1.

To fix this, need to allow colorimetry 1:4:5:1 produced by uvc driver if desired
colorspace is 1:4:7:1. This workaround should be removed later like discussion in
https://gitlab.freedesktop.org/gstreamer/gst-plugins-good/-/merge_requests/969.