v4l2: Use warning log instead of error log and g_warning
Amphion encoder will dequeue one more buffer with bytesused=0 to user as it uses
ringbuffer. Timestamp manager is kept in amphion driver. So this last buffer's
timestamp is marked as last index plus frame->duration, which cannot satisfy
GST_BUFFER_TIMESTAMP (buffer) % GST_SECOND == 0. So use warning log instead.

There is one case that frame[1] and frame[2] have the same timestamp. frame[1]
is decoded but cannot be output, while frame[2] can both be decoded and output.
Then frame list will always has the oldest frame with this timestamp and index
[1] until current frame number becomes 102. At this time, v4l2 decoder drops
frame[1] as (102-1 = 101) >100. This is not a real bug, so use GST_WARNING
instead of g_warning.

Signed-off-by: Hou Qi <qi.hou@nxp.com>